refactor(main): 사다리의 행의 값을 바로 테스트코드에서 설정할 수 있도록 각 Generator들을 함수 인터페이스로 선언
