+++ src/main/java/model/BooleanGenerator.java
@@ -1,17 +1,7 @@
 package model;
 
-import java.util.Random;
+@FunctionalInterface
+public interface BooleanGenerator {
 
-public class BooleanGenerator {
-
-    public boolean generateBoolean() {
-        return new Random().nextBoolean();
-    }
-
-    public boolean updateFalseIfTrue(boolean isTrue) {
-        if (isTrue) {
-            return false;
-        }
-        return new Random().nextBoolean();
-    }
+    boolean updateFalseIfTrue(boolean isTrue);
 }
